Produced appropirate rain radar kml plots using the improved nimrod reader
Here are are the 5 rain radar kml files that correspond to the 5 chosen
catchments. Each file contains the rain radar cells that line up with
the matching catchment. These can be read into Google Earth so that they
overlay with the exsisting kml data for each of the candidate catchments


Former-commit-id: 9a3e6590b58e0d53bf1ca19d293a2da5a093914b